[#1300] Make iceoryx2-bb-testing compatible with no_std
