+++ iceoryx2-bb/testing/Cargo.toml
@@ -19,5 +19,6 @@ std = [
 
 [dependencies]
 iceoryx2-bb-elementary-traits = { workspace = true }
+iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-pal-configuration = { workspace = true }
 iceoryx2-pal-testing = { workspace = true }
